+++ dmenu.c
@@ -6,6 +6,9 @@
 #include <stdio.h>
 #include <errno.h>
 #include <sys/types.h>
+#include <fcntl.h>
+#include <poll.h>
+#include <unistd.h>
 
 #include <tllist.h>
 
@@ -15,15 +18,43 @@
 #include "char32.h"
 
 void
-dmenu_load_entries(struct application_list *applications)
+dmenu_load_entries(struct application_list *applications, int abort_fd)
 {
     tll(char32_t *) entries = tll_init();
 
     char *line = NULL;
     size_t alloc_size = 0;
 
+    int flags;
+    if ((flags = fcntl(STDIN_FILENO, F_GETFL)) < 0 ||
+        fcntl(STDIN_FILENO, F_SETFL, flags | O_NONBLOCK) < 0)
+    {
+        LOG_ERRNO("failed to set O_NONBLOCK on stdin");
+        goto out;
+    }
+
     errno = 0;
     while (true) {
+        struct pollfd fds[] = {
+            {.fd = STDIN_FILENO, .events = POLLIN},
+
+            /* Must be last */
+            {.fd = abort_fd, .events = POLLIN},
+        };
+
+        size_t count = sizeof(fds) / sizeof(fds[0]);
+        if (abort_fd < 0)
+            count--;
+
+        int ret = poll(fds, count, -1);
+        if (ret < 0) {
+            LOG_ERRNO("failed to poll stdin");
+            break;
+        }
+
+        if (fds[1].revents & (POLLIN | POLLHUP))
+            break;
+
         ssize_t len = getline(&line, &alloc_size, stdin);
 
         if (len == -1) {
@@ -45,6 +76,8 @@ dmenu_load_entries(struct application_list *applications)
     }
 
     free(line);
+
+out:
 
     applications->v = calloc(tll_length(entries), sizeof(applications->v[0]));
     applications->count = tll_length(entries);